refactor(init): extract creation flow (#520) (#536)
* refactor(init): extract creation flow (#520)

* refactor(init): extract pr and issue api surface (#539) (#540)

Problem: `init.lua` still owned the remaining PR and issue entrypoints,
which left the API surface split incomplete. The current `#520` stack
point also trips a LuaLS warning in `cmd.lua` during the full CI gate.

Solution: move the PR-facing and issue-facing entrypoints into dedicated
`forge.pr` and `forge.issue` facades while keeping `init.lua` as lazy
re-exports. Predeclare the optional PR resolution locals in `cmd.lua`
so this branch passes the current CI toolchain.

* refactor(cmd): extract dispatch and resolution (#525) (#537)

* refactor(cmd): extract dispatch and resolution (#525)

* refactor(cmd): extract completion engine (#526) (#538)
